Fix review comments

+++ offload/plugins-nextgen/common/include/PluginInterface.h
@@ -794,8 +794,9 @@ struct GenericDeviceTy : public DeviceAllocatorTy {
   /// Get the unique identifier of the device.
   const char *getDeviceUid() const { return DeviceUid.c_str(); }
 
-  /// Get the total shared memory per block that can be used in any kernel.
-  uint32_t getMaxBlockSharedMemSize() const { return MaxBlockSharedMemSize; }
+  /// Get the total shared memory per block (in bytes) that can be used in any
+  /// kernel.
+  size_t getMaxBlockSharedMemSize() const { return MaxBlockSharedMemSize; }
 
   /// Set the context of the device if needed, before calling device-specific
   /// functions. Plugins may implement this function as a no-op if not needed.
@@ -1256,7 +1257,7 @@ protected:
 #endif
 
   /// The total per-block native shared memory that a kernel may use.
-  uint32_t MaxBlockSharedMemSize = 0;
+  size_t MaxBlockSharedMemSize = 0;
 };
 
 /// Class implementing common functionalities of offload plugins. Each plugin
